safari-600.7-branch/LayoutTests/proximity/optional-event-properties-expected.txt

Tests the optional properties of DeviceProximityEvent. Each property should be null if not set, or set to null or undefined.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


event = document.createEvent('DeviceProximityEvent')
PASS event.value == Number.POSITIVE_INFINITY is true
PASS event.min == Number.NEGATIVE_INFINITY is true
PASS event.max == Number.POSITIVE_INFINITY is true
PASS successfullyParsed is true

TEST COMPLETE
